Work around certificate exceptions on Docker/Linux
The user certificate store isn't available on Linux, resulting in an
error that prevents the exposure key set from being published when
running in a Docker container on Linux.

These changes don't fix the problem, but simply work around it by
catching the exceptions and forcing the use of a dummy signature
instead.